+++ src/js/ui/NewDataExploration/LayerHelpers.js
